Progsik/Assignment 1/SecFit/backend/requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Django==4.0.8
django-cors-headers==3.13.0
django-rest-knox==5.0.0
djangorestframework==3.13.1
dynaconf==3.2.5
hvac==2.3.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
loguru==0.7.2
MarkupSafe==2.1.5
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.33.0
win32-setctime==1.1.0
djangorestframework-simplejwt==5.3.1
